feat: 综合更新 - MallCenter优化 + 前后端修复 + 文档清理
MallCenter 小程序:
- 修复 WXSS 组件选择器错误 (goods-card)
- 替换 emoji 为纯中文文字
- 修复 wx:key 警告
- 优化商品列表和详情页样式

MallCenter 后台:
- 更新 API 配置和安全设置
- 优化订单和商品管理

React Native 前端:
- 修复供应商/客户 API 调用 (params 修复)
- 字段名对齐后端 (phone, shippingAddress)
- 优化管理界面

后端 Java:
- 新增 Controller (Disposal, Shipment, Traceability 等)
- 完善 DTO 和 Repository
- AI 分析服务优化

文档清理:
- 移除过期报告和测试文件
- 整理 PRD 文档结构

+++ backend-java/src/main/java/com/cretas/aims/controller/WhitelistController.java
@@ -47,7 +47,7 @@ public class WhitelistController {
         log.info("批量添加白名单: factoryId={}, count={}", factoryId, request.getEntries().size());
         // 获取当前用户ID
         String token = TokenUtils.extractToken(authorization);
-        Integer userId = mobileService.getUserFromToken(token).getId();
+        Long userId = mobileService.getUserFromToken(token).getId();
 
         WhitelistDTO.BatchResult result = whitelistService.batchAdd(factoryId, userId, request);
         return ApiResponse.success(result);
